\name{gsom-package}
\alias{gsom-package}
\alias{gsom}
\docType{package}
\title{
\packageTitle{gsom}
}
\description{
\packageDescription{gsom}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gsom}
\packageIndices{gsom}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gsom}

Maintainer: \packageMaintainer{gsom}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
